+++ packages/eslint-config/index.js
@@ -1,199 +1,209 @@
 module.exports = {
-  parser: "@typescript-eslint/parser",
+  parser: '@typescript-eslint/parser',
   extends: [
-    "plugin:react/recommended",
-    "prettier",
-    "plugin:prettier/recommended",
-    "plugin:tailwindcss/recommended",
-    "plugin:@tanstack/eslint-plugin-query/recommended",
+    'plugin:react/recommended',
+    'prettier',
+    'plugin:prettier/recommended',
+    'plugin:tailwindcss/recommended',
+    'plugin:@tanstack/eslint-plugin-query/recommended'
   ],
-  plugins: [
-    "react",
-    "import-helpers",
-    "react-hooks",
-    "perfectionist",
-    "@tanstack/query",
-    "jest-formatting",
-    "jest",
-  ],
+  plugins: ['react', 'import-helpers', 'react-hooks', 'perfectionist', '@tanstack/query', 'jest-formatting', 'jest', 'testing-library'],
   overrides: [
     {
       env: {
-        node: true,
+        node: true
       },
-      files: ["*.ts", "*.tsx"],
+      files: ['*.ts', '*.tsx'],
       extends: [
-        "plugin:@typescript-eslint/recommended",
-        // 'plugin:@typescript-eslint/recommended-requiring-type-checking',
+        'plugin:@typescript-eslint/recommended'
       ],
       parserOptions: {
-        project: ["./tsconfig.json"],
-      },
-    },
+        project: ['./tsconfig.json']
+      }
+    }
   ],
   rules: {
-    "no-console": "warn",
-    "linebreak-style": "off",
+    'no-console': 'warn',
+    'linebreak-style': 'off',
 
-    "@typescript-eslint/no-unused-vars": "error",
-    "@typescript-eslint/unbound-method": "off",
-    "@typescript-eslint/no-empty-function": "off",
-    "@typescript-eslint/consistent-type-imports": "off",
-    "@typescript-eslint/no-confusing-void-expression": "off",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/strict-boolean-expressions": "off",
-    "@typescript-eslint/no-empty-function": "off",
-    "@typescript-eslint/no-misused-promises": "off",
-    "@typescript-eslint/no-floating-promises": "off",
-    "@typescript-eslint/no-empty-function": "off",
+    '@typescript-eslint/unbound-method': 'off',
+    '@typescript-eslint/no-unused-vars': 'error',
+    '@typescript-eslint/no-empty-function': 'off',
+    '@typescript-eslint/no-empty-function': 'off',
+    '@typescript-eslint/no-empty-function': 'off',
+    '@typescript-eslint/no-misused-promises': 'off',
+    '@typescript-eslint/no-floating-promises': 'off',
+    '@typescript-eslint/consistent-type-imports': 'off',
+    '@typescript-eslint/strict-boolean-expressions': 'off',
+    '@typescript-eslint/no-confusing-void-expression': 'off',
+    '@typescript-eslint/explicit-function-return-type': 'off',
 
-    "react/react-in-jsx-scope": "off",
+    'react/jsx-newline': 'warn',
+    'react/react-in-jsx-scope': 'off',
 
-    "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "warn",
+    'react-hooks/rules-of-hooks': 'error',
+    'react-hooks/exhaustive-deps': 'warn',
 
-    "tailwindcss/no-custom-classname": "off",
+    'tailwindcss/no-custom-classname': 'off',
 
-    "jest/valid-title": "error",
-    "jest/valid-expect": "error",
-    "jest/no-disabled-tests": "warn",
-    "jest/no-duplicate-hooks": "error",
-    "jest/no-identical-title": "error",
-    "jest-formatting/padding-around-all": 2,
-    "jest/require-top-level-describe": "error",
-    "jest/consistent-test-it": ["error", { fn: "it", withinDescribe: "it" }],
+    'jest/valid-title': 'error',
+    'jest/valid-expect': 'error',
+    'jest/no-disabled-tests': 'warn',
+    'jest/no-duplicate-hooks': 'error',
+    'jest/no-identical-title': 'error',
+    'jest-formatting/padding-around-all': 2,
+    'jest/require-top-level-describe': 'error',
+    'jest/consistent-test-it': ['error', { fn: 'it', withinDescribe: 'it' }],
 
-    "prettier/prettier": [
-      "error",
+    'testing-library/prefer-user-event': 'warn',
+    'testing-library/no-debugging-utils': 'warn',
+    'testing-library/await-async-queries': 'warn',
+    'testing-library/prefer-screen-queries': 'warn',
+    'testing-library/prefer-find-by': 'error',
+    'testing-library/no-unnecessary-act': 'error',
+    'testing-library/no-await-sync-events': 'error',
+
+    'padding-line-between-statements': [
+      'error',
+      { prev: '*', next: 'return', blankLine: 'always' },
+      { prev: '*', next: 'function', blankLine: 'always' },
+      { prev: '*', next: 'if', blankLine: 'always' },
+      { next: '*', blankLine: 'always', prev: 'multiline-const' },
+      { next: '*', blankLine: 'always', prev: 'multiline-let' },
+      { next: '*', prev: 'block', blankLine: 'always' },
+      { prev: '*', next: 'block', blankLine: 'always' },
+      { next: '*', prev: 'block-like', blankLine: 'always' },
+      { prev: '*', next: 'block-like', blankLine: 'always' },
+      { next: '*', prev: 'expression', blankLine: 'always' },
+    ],
+
+    'prettier/prettier': [
+      'error',
       {
-        endOfLine: "auto",
+        endOfLine: 'auto',
+        singleQuote: true,
+        arrowParens: 'avoid',
+        bracketSpacing: true,
+        trailingComma: 'all',
+        bracketSameLine: true,
       },
     ],
 
-    // "import-helpers/order-imports": [
-    //   "warn",
-    //   {
-    //     newlinesBetween: "always",
-    //     groups: [
-    //       ["/^react/", "module"],
-    //       "/^~/",
-    //       ["parent", "sibling", "index"],
-    //     ],
-    //     alphabetize: { order: "ignore", ignoreCase: true },
-    //   },
-    // ],
-
-    "perfectionist/sort-object-types": [
-      "error",
+    'perfectionist/sort-maps': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-array-includes": [
-      "error",
+    'perfectionist/sort-enums': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'alphabetical',
       },
     ],
-    "perfectionist/sort-classes": [
-      "error",
+    'perfectionist/sort-classes': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-enums": [
-      "error",
+    'perfectionist/sort-exports': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-exports": [
-      "error",
+    'perfectionist/sort-jsx-props': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-interfaces": [
-      "error",
+    'perfectionist/sort-union-types': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-jsx-props": [
-      "error",
+    'perfectionist/sort-object-types': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-map-elements": [
-      "error",
+    'perfectionist/sort-named-exports': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-named-exports": [
-      "error",
+    'perfectionist/sort-named-imports': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-named-imports": [
-      "error",
+    'perfectionist/sort-array-includes': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
       },
     ],
-    "perfectionist/sort-imports": [
-      "error",
+    'perfectionist/sort-interfaces': [
+      'error',
       {
-        type: "line-length",
-        order: "asc",
+        order: 'asc',
+        type: 'line-length',
+        'partition-by-new-line': true,
+      },
+    ],
+    'perfectionist/sort-object-types': [
+      'error',
+      {
+        order: 'asc',
+        type: 'line-length',
+        'partition-by-new-line': true,
+      },
+    ],
+    'perfectionist/sort-objects': [
+      'error',
+      {
+        order: 'asc',
+        type: 'line-length',
+        'partition-by-comment': true,
+        'partition-by-new-line': true,
+      },
+    ],
+    'perfectionist/sort-imports': [
+      'error',
+      {
+        order: 'asc',
+        type: 'line-length',
+        'internal-pattern': ['@*/**', '@*'],
         groups: [
-          ["type", "builtin", "external", "unknown"],
-          ["internal-type", "internal"],
+          ['type', 'builtin', 'external', 'unknown'],
+          ['internal-type', 'internal'],
           [
-            "parent-type",
-            "sibling-type",
-            "index-type",
-            "parent",
-            "sibling",
-            "index",
-            "object",
+            'parent-type',
+            'sibling-type',
+            'index-type',
+            'parent',
+            'sibling',
+            'index',
+            'object',
           ],
         ],
       },
     ],
-    "perfectionist/sort-object-types": [
-      "error",
-      {
-        type: "line-length",
-        order: "asc",
-      },
-    ],
-    "perfectionist/sort-objects": [
-      "error",
-      {
-        type: "line-length",
-        order: "asc",
-        "partition-by-comment": true,
-      },
-    ],
-    "perfectionist/sort-union-types": [
-      "error",
-      {
-        type: "line-length",
-        order: "asc",
-      },
-    ],
-  },
-};
+  }
+}